Prepare scenes and assets
