+++ package.json
@@ -7,7 +7,8 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "webpack --mode=development",
     "prod": "webpack --mode=production",
-    "start": "webpack serve --open --mode development"
+    "start": "webpack serve --open --mode development",
+    "eslint": "npx eslint . --fix"
   },
   "repository": {
     "type": "git",
