+++ src/config/game_options.js
@@ -1,44 +1,20 @@
 const gameOptions = {
-
-  // platform speed range, in pixels per second
-  platformSpeedRange: [300, 300],
-
-  // mountain speed, in pixels per second
-  mountainSpeed: 80,
-
-  // spawn range, how far should be the rightmost platform from the right edge
-  // before next platform spawns, in pixels
-  spawnRange: [80, 300],
-
-  // platform width range, in pixels
-  platformSizeRange: [90, 300],
-
-  // a height range between rightmost platform and next platform to be spawned
-  platformHeightRange: [-5, 5],
-
-  // a scale to be multiplied by platformHeightRange
-  platformHeighScale: 20,
-
-  // platform max and min height, as screen height ratio
-  platformVerticalLimit: [0.4, 0.8],
-
+  GAME_WIDTH: 1334,
+  GAME_HEIGHT: 750,
   // player gravity
-  playerGravity: 900,
+  PLAYER_GRAVITY: 900,
 
   // player jump force
-  jumpForce: 400,
-
-  // player starting X position
-  playerStartPosition: 200,
+  JUMP_FORCE: 400,
 
   // consecutive jumps allowed
-  jumps: 2,
+  JUMPS: 2,
 
   // % of probability a coin appears on the platform
-  coinPercent: 25,
+  COIN_PERCENT: 25,
 
   // % of probability a fire appears on the platform
-  firePercent: 25,
+  FIRE_PERCENT: 25,
 };
 
 export default gameOptions;
